requirements.txt
Django==2.2.1
djangorestframework==3.9.4
dj-database-url==0.5.0
pytz==2019.1
sqlparse==0.3.0
